requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.4.1
billiard==3.3.0.22
celery==3.1.20
decorator==4.0.9
diff-match-patch==20121119
Django==1.9.3
django-celery==3.1.17
django-extensions==1.6.1
django-import-export==0.4.2
GeoIP==1.3.2
ipython==4.1.2
ipython-genutils==0.1.0
kombu==3.0.33
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
pytz==2015.7
redis==2.10.5
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
tablib==0.11.2
traitlets==4.1.0
